+++ src/trackers/lineTracker.ts
@@ -5,25 +5,25 @@ import { isTextEditor } from './../constants';
 
 export { GitLineState } from './gitDocumentState';
 
-export interface LineChangeEvent {
+export interface LinesChangeEvent {
 
     readonly editor: TextEditor | undefined;
-    readonly line: number | undefined;
+    readonly lines: number[] | undefined;
 
-    readonly reason: 'editor' | 'line';
+    readonly reason: 'editor' | 'lines';
     readonly pending?: boolean;
 }
 
 export class LineTracker<T> extends Disposable {
-    private _onDidChangeActiveLine = new EventEmitter<LineChangeEvent>();
-    get onDidChangeActiveLine(): Event<LineChangeEvent> {
-        return this._onDidChangeActiveLine.event;
+    private _onDidChangeActiveLines = new EventEmitter<LinesChangeEvent>();
+    get onDidChangeActiveLines(): Event<LinesChangeEvent> {
+        return this._onDidChangeActiveLines.event;
     }
 
     private _disposable: Disposable | undefined;
     private _editor: TextEditor | undefined;
 
-    state: T | undefined;
+    private readonly _state: Map<number, T | undefined> = new Map();
 
     constructor() {
         super(() => this.dispose());
@@ -39,34 +39,50 @@ export class LineTracker<T> extends Disposable {
 
         this.reset();
         this._editor = editor;
-        this._line = editor !== undefined ? editor.selection.active.line : undefined;
+        this._lines = editor !== undefined ? editor.selections.map(s => s.active.line) : undefined;
 
-        this.fireLineChanged({ editor: editor, line: this._line, reason: 'editor' });
+        this.fireLinesChanged({ editor: editor, lines: this._lines, reason: 'editor' });
     }
 
     private onTextEditorSelectionChanged(e: TextEditorSelectionChangeEvent) {
         // If this isn't for our cached editor and its not a real editor -- kick out
         if (this._editor !== e.textEditor && !isTextEditor(e.textEditor)) return;
 
-        const reason = this._editor === e.textEditor ? 'line' : 'editor';
+        const reason = this._editor === e.textEditor ? 'lines' : 'editor';
 
-        const line = e.selections[0].active.line;
-        if (this._editor === e.textEditor && this._line === line) return;
+        const lines = e.selections.map(s => s.active.line);
+        if (this._editor === e.textEditor && this.includesAll(lines)) return;
 
         this.reset();
         this._editor = e.textEditor;
-        this._line = line;
+        this._lines = lines;
 
-        this.fireLineChanged({ editor: this._editor, line: this._line, reason: reason });
+        this.fireLinesChanged({ editor: this._editor, lines: this._lines, reason: reason });
     }
 
-    private _line: number | undefined;
-    get line() {
-        return this._line;
+    getState(line: number): T | undefined {
+        return this._state.get(line);
+    }
+
+    setState(line: number, state: T | undefined) {
+        this._state.set(line, state);
+    }
+
+    private _lines: number[] | undefined;
+    get lines(): number[] | undefined {
+        return this._lines;
+    }
+
+    includes(line: number): boolean {
+        return this._lines !== undefined && this._lines.includes(line);
+    }
+
+    includesAll(lines: number[] | undefined): boolean {
+        return LineTracker.includesAll(lines, this._lines);
     }
 
     reset() {
-        this.state = undefined;
+        this._state.clear();
     }
 
     start() {
@@ -83,46 +99,53 @@ export class LineTracker<T> extends Disposable {
     stop() {
         if (this._disposable === undefined) return;
 
-        if (this._lineChangedDebounced !== undefined) {
-            this._lineChangedDebounced.cancel();
+        if (this._linesChangedDebounced !== undefined) {
+            this._linesChangedDebounced.cancel();
         }
 
         this._disposable.dispose();
         this._disposable = undefined;
     }
 
-    private _lineChangedDebounced: (((e: LineChangeEvent) => void) & IDeferrable) | undefined;
+    private _linesChangedDebounced: (((e: LinesChangeEvent) => void) & IDeferrable) | undefined;
 
-    private fireLineChanged(e: LineChangeEvent) {
-        if (e.line === undefined) {
+    private fireLinesChanged(e: LinesChangeEvent) {
+        if (e.lines === undefined) {
             setImmediate(() => {
                 if (window.activeTextEditor !== e.editor) return;
 
-                if (this._lineChangedDebounced !== undefined) {
-                    this._lineChangedDebounced.cancel();
+                if (this._linesChangedDebounced !== undefined) {
+                    this._linesChangedDebounced.cancel();
                 }
 
-                this._onDidChangeActiveLine.fire(e);
+                this._onDidChangeActiveLines.fire(e);
             });
 
             return;
         }
 
-        if (this._lineChangedDebounced === undefined) {
-            this._lineChangedDebounced = Functions.debounce((e: LineChangeEvent) => {
+        if (this._linesChangedDebounced === undefined) {
+            this._linesChangedDebounced = Functions.debounce((e: LinesChangeEvent) => {
                 if (window.activeTextEditor !== e.editor) return;
-                // Make sure we are still on the same line
-                if (e.line !== (e.editor && e.editor.selection.active.line)) return;
+                // Make sure we are still on the same lines
+                if (!LineTracker.includesAll(e.lines , (e.editor && e.editor.selections.map(s => s.active.line)))) return;
 
-                this._onDidChangeActiveLine.fire(e);
+                this._onDidChangeActiveLines.fire(e);
             }, 250, { track: true });
         }
 
         // If we have no pending moves, then fire an immediate pending event, and defer the real event
-        if (!this._lineChangedDebounced.pending!()) {
-            this._onDidChangeActiveLine.fire({ ...e, pending: true });
+        if (!this._linesChangedDebounced.pending!()) {
+            this._onDidChangeActiveLines.fire({ ...e, pending: true });
         }
 
-        this._lineChangedDebounced(e);
+        this._linesChangedDebounced(e);
+    }
+
+    static includesAll(lines1: number[] | undefined, lines2: number[] | undefined): boolean {
+        if (lines1 === undefined && lines2 === undefined) return true;
+        if (lines1 === undefined || lines2 === undefined) return false;
+
+        return lines2.length === lines1.length && lines2.every((v, i) => v === lines1[i]);
     }
 }